Removed shared package statistics-db
