+++ backend/app/statistics-syncer/package.json
@@ -9,7 +9,7 @@
         }
     },
     "scripts": {
-        "build": "tsc --build tsconfig.build.json",
+        "build": "tsc --build tsconfig.build.json && mkdir -p ./dist/src/migrations && rsync --delete --exclude='*.ts' --exclude='*.js' --exclude='*.map' -r --checksum ./src/migrations/ ./dist/src/migrations/",
         "build:full": "yarn -s clear && yarn -s build",
         "clear": "rm -rf ./dist && rm -f *.tsbuildinfo",
         "dev": "wait-on ../../shared/middleware/dist/index.js && concurrently -r 'yarn -s build --watch --preserveWatchOutput' \"wait-on ./dist/index.js && nodemon --quiet --inspect=5859 --watch dist --watch '../../../shared/*/dist/' --watch '../../shared/*/dist/' --ext .ts,.json,.sql,.js --watch .env.json --delay 1000ms --exec 'node --enable-source-maps ./dist/index.js' --signal SIGTERM\"",
@@ -33,7 +33,6 @@
         "@stamhoofd/models": "2.140.0",
         "@stamhoofd/queues": "2.140.0",
         "@stamhoofd/sql": "2.140.0",
-        "@stamhoofd/statistics-db": "2.140.0",
         "@stamhoofd/structures": "2.140.0",
         "@stamhoofd/types": "2.140.0",
         "@stamhoofd/utility": "2.140.0",
