--- backend/shared/statistics-db/stamhoofd.d.ts
@@ -1,17 +1,0 @@
-import '@stamhoofd/types/global';
-import type { StatisticsEnvironment } from '@stamhoofd/types/Environment';
-
-export {};
-
-/**
- * Stamhoofd uses a global variable to store some configurations. We don't use process.env because we can only store
- * strings into those files. And we need objects for our localized domains (different domains for each locale).
- * Having to encode and decode those values would be inefficient.
- *
- * So we use our own global configuration variable: STAMHOOFD. Available everywhere and contains
- * other information depending on the environment (frontend/backend/shared). TypeScript will
- * always suggest the possible keys.
- */
-declare global {
-    const STAMHOOFD: StatisticsEnvironment;
-}
