+++ package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "build:backend:apps": "lerna run build --scope=@stamhoofd/backend --scope=@stamhoofd/backend-renderer --scope=@stamhoofd/backend-backup --scope=@stamhoofd/backend-redirecter --scope=@stamhoofd/backend-statistics-syncer",
-    "build:backend:shared": "echo '\\033[35m[BUILD]\\033[0m Building shared backend dependencies...' && cd backend/shared/queues && yarn -s build && cd ../env && yarn -s build && cd ../i18n && yarn -s build && cd ../sql && yarn -s build && cd ../email && yarn -s build && cd ../models && yarn -s build && cd ../logging && yarn -s build && cd ../crons && yarn -s build && cd ../middleware && yarn -s build && cd ../statistics-db && yarn -s build && cd ../.. && echo '\\033[35m[BUILD]\\033[0m Done building backend shared dependencies.'",
+    "build:backend:shared": "echo '\\033[35m[BUILD]\\033[0m Building shared backend dependencies...' && cd backend/shared/queues && yarn -s build && cd ../env && yarn -s build && cd ../i18n && yarn -s build && cd ../sql && yarn -s build && cd ../email && yarn -s build && cd ../models && yarn -s build && cd ../logging && yarn -s build && cd ../crons && yarn -s build && cd ../middleware && yarn -s build && cd ../.. && echo '\\033[35m[BUILD]\\033[0m Done building backend shared dependencies.'",
     "build:global:shared": "echo '\\033[35m[BUILD]\\033[0m Building globally shared dependencies...' && cd shared/types && yarn -s build && cd ../test-utils && yarn -s build && cd ../utility && yarn -s build && cd ../metabase && yarn -s build && cd ../cli && yarn -s build && cd ../excel-writer && yarn -s build && cd ../structures && yarn -s build && cd ../sgv && yarn -s build && cd ../object-differ && yarn -s build && cd ../locales && yarn -s build && cd ../.. && echo '\\033[35m[BUILD]\\033[0m Done building globally shared dependencies.'",
     "build:shared": "yarn -s build:global:shared && yarn -s build:backend:shared",
     "build:shared:watch": "npx nodemon --quiet --signal SIGTERM --watch shared --watch backend/shared --ignore './shared/*/dist/' --ignore './backend/shared/*/dist/' --ext .ts,.json,.sql,.mjs --exec 'yarn -s build:shared || exit 0'",
